Rename `SuperMechs.types` to `game_types`
